requirements.txt
Pyrogram==2.0.30
TgCrypto==1.2.3
aiohttp
python-dotenv
beautifulsoup4==4.11.1
requests==2.28.1
aiofiles
translators
motor==2.5.1
pymongo[srv]==3.12.3
python-slugify